Proof of concept for holistic approach
This is a very rough sketch of what this might want to
look like, longer term. Very few of these decisions are
clearly correct (for instance using the Matcher trait
instead of just returning
`impl Fn(usize) -> Result<&str, Error>`, or something)
but it offers an outline of how we might approach
this problem in a hacky but reasonably structured way?